Implemented GUI
Switched out JPanel Array with a JButton Array, added action listener
to each index in the array. When user clicks on first piece it changes
the background of the piece they want to move. I have also added a
message that pops that lets the user know they won or lost.